+++ CHANGES.rst
@@ -12,6 +12,10 @@ Changes
 
 Changes:
 --------
+- Add `HTML` shortcut links to easily toggle contextually between full `Provider` or only local `Process` listing.
+- Add `HTML` support of ``/processes/?providers=true`` to list all `Provider` and local `Process` references.
+- Add ``?provider={providerID}`` and ``?service={providerID}`` query parameters to ``/processes/{processID}`` request
+  as alias to resolve the corresponding ``/providers/{providerID}/processes/{processID}`` resource.
 - Add links and breadcrumbs including the `Process` ID within `HTML` pages of `Job` responses
   if referenced by the request.
 - Add the full URI definitions for ``Accept-Profile`` header that correspond to equivalent shorthand notation